% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzywuzzy.R
\name{compute_fuzzy_scores}
\alias{compute_fuzzy_scores}
\title{Add Partial Token Sort_Ratio Score}
\usage{
compute_fuzzy_scores(df, s1, s2)
}
\arguments{
\item{df}{data.frame}

\item{s1}{first string}

\item{s2}{second string}
}
\value{
data.frame
}
\description{
=> add a SCORE column to the data.frame using
  v_fz_tk_sort_r method
}
\author{
Marc Lamarine
}
\keyword{internal}
