% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzywuzzy.R
\name{fz_tk_sort_r}
\alias{fz_tk_sort_r}
\title{Partial_token_sort_ratio 
 
 => wrapper to get the Partial_token_sort_ratio function}
\usage{
fz_tk_sort_r(init = .get_fuzz_matcher())
}
\arguments{
\item{init}{(a FuzzMatcher Object)}
}
\value{
an integer
}
\description{
Partial_token_sort_ratio 
 
 => wrapper to get the Partial_token_sort_ratio function
}
\author{
Marc Lamarine
}
