% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzywuzzy.R
\name{v_fz_tk_sort_r}
\alias{v_fz_tk_sort_r}
\title{Vectorized Partial Token Sort_Ratio 
 
 => compute the parallel partial token sort ratio between two strings}
\usage{
v_fz_tk_sort_r(string1 = NULL, string2 = NULL, force_ascii = TRUE,
  full_process = TRUE)
}
\arguments{
\item{string1}{a character string.}

\item{string2}{a character string.}

\item{force_ascii}{allow only ASCII characters (force convert to ascii)}

\item{full_process}{either TRUE or FALSE. If TRUE then it process the string by : 1. removing all  but letters and numbers, 2. trim whitespace, 3. force to lower case}
}
\value{
a vector
}
\description{
Vectorized Partial Token Sort_Ratio 
 
 => compute the parallel partial token sort ratio between two strings
}
\author{
Marc Lamarine
}
